third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/error-conditions.html

<!DOCTYPE html>
<title>Tests that error conditions cause exceptions</title>
<link rel="help" href="https://github.com/WICG/html-in-canvas">
<link rel="author" href="mailto:chrishtr@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
#child {
  width: 100px;
  height: 100px;
  background: green;
}
#canvas {
  background: grey;
}
</style>

<canvas id=canvas width="200" height="200" layoutsubtree>
  <div id=child>
    <div id=grandchild></div>
  </div>
</canvas>
<canvas id=canvaswithoutlayout layoutsubtree>
  <div id=child2></div>
</div>

</canvas>

<script>
promise_test(async () => {
  assert_throws_js(TypeError,
    () => canvas.getContext("2d").drawElementImage(grandchild, 20, 30),
    "Can't draw non-direct children.");

  canvaswithoutlayout.removeAttribute("layoutsubtree");
  assert_throws_js(TypeError,
    () => canvaswithoutlayout.getContext("2d").drawElementImage(
      child2, 20, 30),
    "Can't draw canvas children if layoutsubtree is not specified.");
});
</script>
</html>
